WIP -- use WheelIndicator to get swipe info
Look at the main page with console open.
Need to figure out how to cancel scroll events when necessary and do the
scroll-to-next functionality ourselves.
Need to have a timeout where we ignore additional scroll of
wheelindicator events because sometimes when doing quick left/right
swipes I see false positives in the console.